Consolidate PID checking code in utils package
