+++ src/confab/confab/confab_test.go
@@ -11,6 +11,8 @@ import (
 	"path/filepath"
 	"time"
 
+	"github.com/cloudfoundry-incubator/consul-release/src/confab/utils"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -126,7 +128,7 @@ var _ = Describe("confab", func() {
 
 			pid, err := getPID(pidFile.Name())
 			Expect(err).NotTo(HaveOccurred())
-			Expect(isPIDRunning(pid)).To(BeTrue())
+			Expect(utils.IsPIDRunning(pid)).To(BeTrue())
 
 			stop := exec.Command(pathToConfab,
 				"stop",
@@ -135,7 +137,7 @@ var _ = Describe("confab", func() {
 			Eventually(stop.Run, COMMAND_TIMEOUT, COMMAND_TIMEOUT).Should(Succeed())
 
 			Eventually(func() bool {
-				return isPIDRunning(pid)
+				return utils.IsPIDRunning(pid)
 			}, COMMAND_TIMEOUT, time.Millisecond*250).Should(BeFalse())
 
 			Expect(fakeAgentOutput(consulConfigDir)).To(Equal(FakeAgentOutputData{
@@ -266,7 +268,7 @@ var _ = Describe("confab", func() {
 
 				pid, err := getPID(pidFile.Name())
 				Expect(err).NotTo(HaveOccurred())
-				Expect(isPIDRunning(pid)).To(BeTrue())
+				Expect(utils.IsPIDRunning(pid)).To(BeTrue())
 
 				Expect(fakeAgentOutput(consulConfigDir)).To(Equal(FakeAgentOutputData{
 					PID: pid,
@@ -314,7 +316,7 @@ var _ = Describe("confab", func() {
 
 				pid, err := getPID(pidFile.Name())
 				Expect(err).NotTo(HaveOccurred())
-				Expect(isPIDRunning(pid)).To(BeTrue())
+				Expect(utils.IsPIDRunning(pid)).To(BeTrue())
 
 				Eventually(func() (FakeAgentOutputData, error) {
 					return fakeAgentOutput(consulConfigDir)
@@ -427,7 +429,7 @@ var _ = Describe("confab", func() {
 			Eventually(cmd.Run, COMMAND_TIMEOUT, COMMAND_TIMEOUT).Should(Succeed())
 
 			Eventually(func() bool {
-				return pidIsForRunningProcess(pidFile.Name())
+				return utils.IsRunningProcess(pidFile.Name())
 			}, "5s").Should(BeFalse())
 
 			Eventually(func() (FakeAgentOutputData, error) {
@@ -647,7 +649,7 @@ var _ = Describe("confab", func() {
 
 				pid, err := getPID(pidFile.Name())
 				Expect(err).NotTo(HaveOccurred())
-				Expect(isPIDRunning(pid)).To(BeTrue())
+				Expect(utils.IsPIDRunning(pid)).To(BeTrue())
 			})
 		})
