+++ src/confab/utils/utils.go
@@ -1,7 +1,20 @@
 package utils
 
-import "os"
+import (
+	"io/ioutil"
+	"strconv"
+)
 
-func CheckProcessRunning(process *os.Process) error {
-	return checkProcessRunning(process)
+func IsRunningProcess(pidFilePath string) bool {
+	pidFileContents, err := ioutil.ReadFile(pidFilePath)
+	if err != nil {
+		return false
+	}
+
+	pid, err := strconv.Atoi(string(pidFileContents))
+	if err != nil {
+		return false
+	}
+
+	return IsPIDRunning(pid)
 }
