+++ src/confab/utils/utils_test.go
@@ -1,28 +1,55 @@
 package utils_test
 
 import (
-	"os/exec"
-	"runtime"
+	"io/ioutil"
+	"os"
+	"strconv"
 
 	"github.com/cloudfoundry-incubator/consul-release/src/confab/utils"
 
 	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("a process that is about to die", func() {
-	It("is running, and eventually stops", func() {
-		var cmd *exec.Cmd
-		if runtime.GOOS == "windows" {
-			cmd = exec.Command("ping", "-t", "127.0.0.1")
-		} else {
-			cmd = exec.Command("ping", "127.0.0.1")
-		}
-		Expect(cmd.Start()).To(Succeed())
-		defer cmd.Process.Kill()
-		Expect(utils.CheckProcessRunning(cmd.Process)).To(Succeed())
-		cmd.Process.Kill()
-		cmd.Wait()
-		Expect(utils.CheckProcessRunning(cmd.Process)).ToNot(Succeed())
+var _ = Describe("utils", func() {
+	Describe("IsRunningProcess", func() {
+		var (
+			pidFile *os.File
+		)
+
+		BeforeEach(func() {
+			var err error
+
+			pidFile, err = ioutil.TempFile("", "")
+			Expect(err).NotTo(HaveOccurred())
+		})
+
+		DescribeTable("when the pidfile exists",
+			func(pid string, isRunning bool) {
+				err := ioutil.WriteFile(pidFile.Name(), []byte(pid), os.ModePerm)
+				Expect(err).NotTo(HaveOccurred())
+
+				processIsRunning := utils.IsRunningProcess(pidFile.Name())
+				Expect(processIsRunning).To(Equal(isRunning))
+			},
+			Entry("returns false if the process is not running", "-1", false),
+			Entry("returns true if the process is running", strconv.Itoa(os.Getpid()), true),
+			Entry("returns false if the pidfile contains garbage", "something-bad", false),
+		)
+
+		It("returns false if the pidfile does not exist", func() {
+			processIsRunning := utils.IsRunningProcess("/nonexistent/pidfile")
+			Expect(processIsRunning).To(BeFalse())
+		})
 	})
+
+	DescribeTable("IsPIDRunning",
+		func(pid int, expected bool) {
+			isRunning := utils.IsPIDRunning(pid)
+			Expect(isRunning).To(Equal(expected))
+		},
+		Entry("returns true when the process is running", os.Getpid(), true),
+		Entry("returns false when the process is not running", -1, false),
+	)
 })
