+++ src/confab/utils/utils_unix.go
@@ -7,6 +7,9 @@ import (
 	"syscall"
 )
 
-func checkProcessRunning(process *os.Process) error {
-	return process.Signal(syscall.Signal(0))
+func IsPIDRunning(pid int) bool {
+	// On Unix FindProcess always returns
+	// a non-nil Process and a nil error.
+	process, _ := os.FindProcess(pid)
+	return process.Signal(syscall.Signal(0)) == nil
 }
